# public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: IAO 2026 Registration API
  version: 1.0.0
  description: 'IAO 2026 (XXX Xalqaro astronomiya olimpiadasi, Andijon, 6–14 dekabr 2026) ro''yxatdan o''tish API''si.


    **Oqim:** `register` → emailni tasdiqlash → `participants` qo''shish → `team/submit` (yakuniy yuborish; jamoa qulflanadi).


    **Autentifikatsiya:** `Authorization: Bearer <token>` (Laravel Sanctum). Barcha so''rovlarda `Accept: application/json` yuboring.


    **Umumiy xatolar:** 401 — token yo''q; 403 — ruxsat yo''q; 404 — topilmadi (boshqa jamoa ma''lumoti ham 404); 409 — email tasdiqlanmagan;

    422 — validatsiya; 423 — jamoa yuborilgan yoki muddat tugagan; 429 — juda ko''p so''rov.


    **Sanalar:** `KK.OO.YYYY`. **Fayl yuklash:** `multipart/form-data`. Fayl bilan *yangilash* uchun `POST` yuboring va `_method=PATCH` maydonini
    qo''shing

    (PHP multipart PATCH/PUT ni o''qimaydi). Faylsiz yangilash uchun oddiy JSON `PATCH` ishlaydi.'
servers:
- url: '{baseUrl}/api'
  variables:
    baseUrl:
      default: http://localhost:8000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '`register` yoki `login` qaytargan token'
  schemas:
...
    Participant:
      type: object
      properties:
        id:
          type: integer
          example: 1
        status:
          type: string
          enum: &id001
          - team_leader
          - team_leader_jury
          - observer
          - student
          description: Maqomi
        student_group:
          type: string
          enum: &id002
          - alpha
          - beta
          - gamma
          nullable: true
          description: Guruh. `status=student` bo'lsa majburiy
        previous_prizewinner:
          type: boolean
          default: false
          description: Oldingi IAO I–II diplom sovrindori
        needs_visa_invitation:
          type: boolean
          default: false
          description: Viza taklifnomasi kerakmi
        is_paid:
          type: boolean
          default: false
          readOnly: true
          description: To'lov qildi / qilmadi. Faqat administrator o'zgartiradi (`PATCH /admin/participants/{id}/payment`); jamoa yuborsa e'tiborga olinmaydi.
        paid_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: To'langan deb belgilangan vaqt (ISO 8601)
        payment_note:
          type: string
          nullable: true
          maxLength: 255
          readOnly: true
          description: Tashkilotchilar izohi (kvitansiya raqami va h.k.)
        payment_amount:
          type: number
          format: double
          nullable: true
          readOnly: true
          example: 150
          description: To'lov summasi
        payment_currency:
          type: string
          nullable: true
          maxLength: 3
          readOnly: true
          example: USD
          description: Valyuta (USD, EUR, UZS, RUB). Summa bor, valyuta yo'q bo'lsa — USD
        family_name_en:
          type: string
          description: 1.1 Familiya (inglizcha). Faqat 26 ta lotin harfi; faqat BOSH HARFLARDA yozish rad etiladi
          maxLength: 100
          example: Okello
        first_name_en:
          type: string
          description: 1.2 Ism (inglizcha). Faqat 26 ta lotin harfi; faqat BOSH HARFLARDA yozish rad etiladi
          maxLength: 100
          example: Peter James
        family_name_native:
          type: string
          maxLength: 100
          example: Окелло
          description: 2.1 Familiya (ona tilida, istalgan yozuv)
        first_name_native:
          type: string
          maxLength: 100
          example: Питер Джеймс
          description: 2.2 Ism (ona tilida)
        birth_date:
          type: string
          pattern: ^\d{2}\.\d{2}\.\d{4}$
          example: 12.05.2009
          description: 3 Tug'ilgan sana, KK.OO.YYYY (bugundan oldin)
        birth_place:
          type: string
          maxLength: 255
          example: Kampala, Uganda
          description: 4 Tug'ilgan joy
        sex:
          type: string
          enum: &id003
          - male
          - female
          description: 5 Jinsi
        citizenship:
          type: string
          example: Ugandan
          description: 6.1 Hozirgi fuqarolik
        other_citizenships:
          type: string
          nullable: true
          description: 6.2 Avvalgi/qo'shimcha fuqaroliklar
        ethnicity:
          type: string
          nullable: true
          description: 6.3 Millati
        previous_visits_uz:
          type: string
          nullable: true
          maxLength: 1000
          description: 6.4 O'zbekistonga avvalgi tashriflar
        passport_number:
          type: string
          nullable: true
          maxLength: 50
          example: B1234567
          description: 8.1 Pasport raqami (butun tizimda noyob). Faqat `needs_visa_invitation=true` bo'lsa majburiy.
        passport_issue_date:
          type: string
          pattern: ^\d{2}\.\d{2}\.\d{4}$
          example: 12.05.2009
          nullable: true
          description: 8.2 Berilgan sana (bugun yoki oldin). Faqat `needs_visa_invitation=true` bo'lsa majburiy.
        passport_expiry_date:
          type: string
          pattern: ^\d{2}\.\d{2}\.\d{4}$
          example: 12.05.2009
          nullable: true
          description: 8.3 Amal qilish muddati (kelajakda). Faqat `needs_visa_invitation=true` bo'lsa majburiy.
        passport_issued_by:
          type: string
          nullable: true
          maxLength: 255
          description: 8.4 Pasportni bergan organ. Faqat `needs_visa_invitation=true` bo'lsa majburiy.
        has_passport_scan:
          type: boolean
        has_face_photo:
          type: boolean
        position:
          type: string
          maxLength: 255
          example: Student
          description: 9 Lavozimi (talaba, o'qituvchi...)
        org_name:
          type: string
          maxLength: 255
          description: 10.1 Ish/o'qish tashkilotining nomi
        org_location:
          type: string
          maxLength: 255
          description: 10.2 Tashkilot joylashgan joy
        org_address:
          type: string
          nullable: true
          maxLength: 500
          description: 10.3 Tashkilot manzili (indeks bilan). Faqat `needs_visa_invitation=true` bo'lsa majburiy.
        org_contacts:
          type: string
          nullable: true
          maxLength: 500
          description: 10.4 Faks, telefon, email. Faqat `needs_visa_invitation=true` bo'lsa majburiy.
        graduation_date:
          type: string
          nullable: true
          maxLength: 50
          example: June 2027
          description: 10.5 Tamomlash oy/yili
        previous_olympiads:
          type: string
          nullable: true
          maxLength: 1000
          example: 'IAO-2025: Diploma of participating'
          description: 10.6 Avval qatnashgan IAO/APAO
        home_location:
          type: string
          maxLength: 255
          description: 11.1 Yashash joyi
        home_address:
          type: string
          maxLength: 500
          description: 11.2 Doimiy manzil
        home_phone:
          type: string
          nullable: true
          maxLength: 100
          description: 11.3 Uy telefoni
        mobile_phone:
          type: string
          maxLength: 100
          example: '+256700000000'
          description: 11.4 Mobil telefon
        email:
          type: string
          format: email
          description: 11.5 Email
        official_language:
          type: string
          enum: &id004
          - russian
          - english
          - both
          description: 12.1 Rasmiy til
        native_languages:
          type: string
          maxLength: 255
          example: Luganda
          description: '12.2 Ona tillari (rahbar uchun: tarjima qila oladigan tillar)'
        diet:
          type: string
          enum: &id005
          - standard
          - vegetarian
          - avoid_pork
          default: standard
          description: 13.1 Ovqatlanish
        food_notes:
          type: string
          nullable: true
          maxLength: 1000
          description: 13.2 Ovqat bo'yicha izoh
        medical_notes:
          type: string
          nullable: true
          maxLength: 1000
          description: 13.3 Tibbiy izoh
        tshirt_size:
          type: string
          nullable: true
          enum: &id006
          - XXS
          - XS
          - S
          - M
          - L
          - XL
          - XXL
          - XXXL
          - null
          description: 13.4 Futbolka o'lchami
        emergency_contact:
          type: object
          properties:
            family_name:
              type: string
            first_name:
              type: string
            relation:
              type: string
            age:
              type: integer
              nullable: true
            languages:
              type: string
              nullable: true
            phones:
              type: string
            email:
              type: string
              nullable: true
            telegram:
              type: string
              nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      example:
        message: The family name en field format is invalid.
        errors:
          family_name_en:
          - Faqat ingliz alifbosining 26 ta harfidan foydalaning (The family name en must use Latin letters only).
...
  responses:
    Unauthorized:
      description: Token yo'q yoki yaroqsiz
      content:
        application/json:
          schema: *id007
          example:
            message: Unauthenticated.
...
    NotFound:
      description: Topilmadi (yoki boshqa jamoaga tegishli)
      content:
        application/json:
          schema: *id007
    Unverified:
      description: Email tasdiqlanmagan
      content:
        application/json:
          schema: *id007
          example:
            message: Your email address is not verified.
    Validation:
      description: Validatsiya xatosi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Locked:
      description: Jamoa yakuniy yuborilgan yoki ro'yxatdan o'tish muddati tugagan
      content:
        application/json:
          schema: *id007
          example:
            message: Ro'yxat yakuniy yuborilgan, o'zgartirib bo'lmaydi. Kerak bo'lsa tashkilotchilarga murojaat qiling.
...
paths:
...
  /participants/{id}:
    get:
      tags:
      - Participants
      summary: Bitta ishtirokchi
      security: *id008
      parameters:
      - *id009
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Participant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Participants
      summary: Qisman yangilash
      description: '`PUT` ham xuddi shunday ishlaydi. Faqat yuborilgan maydonlar tekshiriladi.'
      security: *id008
      parameters:
      - *id009
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Faqat o'zgaradigan maydonlar
              properties: *id010
          multipart/form-data:
            schema:
              type: object
              properties:
                status: *id011
                student_group: *id012
                previous_prizewinner: *id013
                needs_visa_invitation: *id014
                family_name_en: *id015
                first_name_en: *id016
                family_name_native: *id017
                first_name_native: *id018
                birth_date: *id019
                birth_place: *id020
                sex: *id021
                citizenship: *id022
                other_citizenships: *id023
                ethnicity: *id024
                previous_visits_uz: *id025
                passport_number: *id026
                passport_issue_date: *id027
                passport_expiry_date: *id028
                passport_issued_by: *id029
                position: *id030
                org_name: *id031
                org_location: *id032
                org_address: *id033
                org_contacts: *id034
                graduation_date: *id035
                previous_olympiads: *id036
                home_location: *id037
                home_address: *id038
                home_phone: *id039
                mobile_phone: *id040
                email: *id041
                official_language: *id042
                native_languages: *id043
                diet: *id044
                food_notes: *id045
                medical_notes: *id046
                tshirt_size: *id047
                emergency_family_name: *id048
                emergency_first_name: *id049
                emergency_relation: *id050
                emergency_age: *id051
                emergency_languages: *id052
                emergency_phones: *id053
                emergency_email: *id054
                emergency_telegram: *id055
                passport_scan: *id056
                face_photo: *id057
                _method:
                  type: string
                  enum:
                  - PATCH
                  description: 'Fayl bilan yangilashda: `POST` + `_method=PATCH`'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Participant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Unverified'
        '422':
          $ref: '#/components/responses/Validation'
        '423':
          $ref: '#/components/responses/Locked'
    delete:
      tags:
      - Participants
      summary: O'chirish (yuklangan fayllari bilan)
      security: *id008
      parameters:
      - *id009
      responses:
        '204':
          description: O'chirildi
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Unverified'
        '423':
          $ref: '#/components/responses/Locked'
...